enable: node -> prom -> alert-manager -> slack

+++ node/install_node_exporter.sh
@@ -11,21 +11,21 @@ export LATEST_TAG=$(curl -v https://github.com/prometheus/node_exporter/releases
 export LATEST_URL="https://github.com/prometheus/node_exporter/releases/download/v${LATEST_TAG}/node_exporter-${LATEST_TAG}.${OS}-${ARCH}.tar.gz"
 export SHA_URL="https://github.com/prometheus/node_exporter/releases/download/v${LATEST_TAG}/sha256sums.txt"
 
-rm -rf /tmp/nodeexpdownload
-mkdir -p /tmp/nodeexpdownload
-pushd /tmp/nodeexpdownload
-wget $LATEST_URL
-wget $SHA_URL
+if [ -z "$(ls ~/.local/bin/node_exporter_${LATEST_TAG})" ]; then
+    rm -rf /tmp/nodeexpdownload
+    mkdir -p /tmp/nodeexpdownload
+    pushd /tmp/nodeexpdownload
+    wget $LATEST_URL
+    wget $SHA_URL
 
-echo "CHECKSUM"
-grep "$(sha256sum node_exporter-*.tar.gz |awk '{print $1}')" sha256sums.txt
+    echo "CHECKSUM"
+    grep "$(sha256sum node_exporter-*.tar.gz |awk '{print $1}')" sha256sums.txt
 
-tar xvf node_exporter-*.tar.gz
-mv node_exporter-*/node_exporter ~/.local/bin/node_exporter_${LATEST_TAG}
-rm -rf ~/.local/bin/node_exporter
-ln -s ~/.local/bin/node_exporter_${LATEST_TAG} ~/.local/bin/node_exporter
+    tar xvf node_exporter-*.tar.gz
+    mv node_exporter-*/node_exporter ~/.local/bin/node_exporter_${LATEST_TAG}
+    rm -rf ~/.local/bin/node_exporter
+    ln -s ~/.local/bin/node_exporter_${LATEST_TAG} ~/.local/bin/node_exporter
+    popd
+fi
 
 ls -al ~/.local/bin
-
-
-popd
